pyproject.toml
[tool.poetry]
name = "mlmonitor"
version = "0.1.0"
description = "Monitoreo automatizado para modelos de scoring de crédito y cobranza"
authors = ["Equipo Analytics <analytics@mlmonitor.local>"]
packages = [{ include = "mlmonitor", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0"
pandas = "^2.2"
numpy = "^1.26"
jinja2 = "^3.1"
weasyprint = "60.2"
pydyf = "0.8.0"
google-cloud-aiplatform = { version = "^1.60", optional = true }
pydantic-settings = "^2.0"
python-dotenv = "^1.0"
scipy = "^1.12"
google-genai = "^1.64.0"

[tool.poetry.extras]
vertex = ["google-cloud-aiplatform"]
oracle = ["python-oracledb"]
aws = ["boto3"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
